Get the rest of the projects over to gyp.  Cleanup in the next cl.

+++ src/afx2lg/test/axefx_test.cc
@@ -70,8 +70,8 @@ TEST_F(AxeFxII, ParseFw9bBankFile) {
     shared_ptr<BlockParameters> amp1 = p.LookupBlock(BLOCK_AMP_1);
     if (amp1) {
       printf("Preset: %i %s\n", (*i).first, p.name().c_str());
-      int amp_id = amp1->GetParamValue(DISTORT_TYPE,
-                                       amp1->active_config() == CONFIG_X);
+      uint16_t amp_id = amp1->GetParamValue(DISTORT_TYPE,
+          amp1->active_config() == CONFIG_X);
       if (amp1->global_block_index())
         printf(" - global=%i\n", amp1->global_block_index());
       printf(" - amp=%i (guessing='%s')\n", amp_id, GetAmpName(amp_id));
@@ -247,8 +247,8 @@ TEST_F(AxeFxII, ParseSystemBackup) {
   EXPECT_EQ(710, global_blocks.size());
   shared_ptr<BlockParameters> amp1 = global_blocks[0];
   EXPECT_EQ(BLOCK_AMP_1, amp1->block());
-  int amp_id_x = amp1->GetParamValue(DISTORT_TYPE, true);
-  int amp_id_y = amp1->GetParamValue(DISTORT_TYPE, false);
+  uint16_t amp_id_x = amp1->GetParamValue(DISTORT_TYPE, true);
+  uint16_t amp_id_y = amp1->GetParamValue(DISTORT_TYPE, false);
 #if defined(_DEBUG)
   if (amp1->global_block_index())
     std::cout << " - global=" << amp1->global_block_index() << std::endl;
